+++ lib/QECCore/ext/QECCoreNemoExt/QECCoreNemoExt.jl
@@ -4,9 +4,14 @@ using QECCore
 using DocStringExtensions
 
 import Nemo
-import Nemo: GF, matrix, rank, transpose
+import Nemo: GF, matrix, rank, transpose, finite_field, GF, polynomial_ring, evaluate,
+    FqFieldElem, FqPolyRingElem, degree, is_irreducible, gcd, derivative, matrix, inv,
+    is_zero, coeff, is_monic, one
 
-import QECCore: code_k, parity_matrix_x, parity_matrix_z
+import QECCore: code_k, parity_matrix_x, parity_matrix_z, parity_matrix, generator_polynomial
+
+import Random
+import Random: MersenneTwister
 
 function QECCore.code_k(c::AbstractCSSCode)
     n = code_n(c)
@@ -23,5 +28,7 @@ function QECCore.code_k(c::AbstractCECC)
     rank_H = rank(matrix(F₂, H))
     return n - rank_H
 end
+
+include("goppa.jl")
 
 end # module