+++ test/test_ecc_codeproperties.jl
@@ -1,4 +1,4 @@
-@testitem "ECC code properties" tags=[:ecc] begin
+@testitem "ECC code properties" tags=[:ecc, :ecc_code_properties] begin
     using QuantumClifford.ECC
     using QuantumClifford.ECC: AbstractECC
 
@@ -19,14 +19,14 @@
     end
 
     @testset "is CSS" begin
-        for code in all_testablable_code_instances()
+        for code in all_testable_code_instances()
             H = parity_checks(code)
             @test iscss(code) in (is_css_matrix(H), nothing)
         end
     end
 
     @testset "code tableau consistency" begin
-        for code in all_testablable_code_instances()
+        for code in all_testable_code_instances()
             H = parity_checks(code)
             @test nqubits(code) == size(H, 2) == code_n(code)
             @test size(H, 1) == code_s(code)
@@ -35,5 +35,12 @@
             @test rank <= size(H, 1)
             @test QuantumClifford.stab_looks_good(copy(H), remove_redundant_rows=true)
         end
+    end
+
+    @testset "is degenerate function - test on popular codes" begin
+        @test isdegenerate(Shor9()) == true
+        @test isdegenerate(Steane7()) == false
+        @test isdegenerate(Steane7(), 2) == true
+        @test isdegenerate(Bitflip3()) == true
     end
 end